Fixes method calls to update on show objects
Previously, calling update on an object of class Camdram::Show resulted in that classes initialize method being bypassed and its instance varibales being set directly
from the JSON hash. This meant that the society, venue and performance instance variables were incorrectly set as they should reference the appropriate objects, not a
Ruby hash or array.

This commit fixes that behaviour.