Send access_token as header rather than query string
Previously, the API access token obtained through an OAuth login was submitted to the Camdram API along with each request by appending a query string. It is possible that
this method is insecure due to the query strings appearing in server access logs, browser histories, Referrer headers or a combination thereof.

More information is available at https://blog.httpwatch.com/2009/02/20/how-secure-are-query-strings-over-https

This commit submits the Camdram API access token as a header sent with the HTTP request which helps to mitigate some of the security concerns touched upon previously.